feat(config-sequence): extract each phase config to a new component
feat(config-sequence): extract ResponsePhase config

feat(config-sequence): extract ConfrontingViewPhase config

feat(config-sequence): extract ResultPhase config

chore(config-sequence): refactor folder
