+++ ui-components/src/stories/sequence/configuration/phase/ResponsePhaseConfiguration.stories.ts
@@ -1,0 +1,68 @@
+/*
+ * Elaastic - formative assessment system
+ * Copyright (C) 2019. University Toulouse 1 Capitole, University Toulouse 3 Paul Sabatier
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+import type {Meta, StoryObj} from '@storybook/vue3-vite';
+import ResponsePhaseConfiguration from "@/components/sequence/configuration/phase/ResponsePhaseConfiguration.vue";
+import {fn} from "storybook/test";
+
+
+// More on how to set up stories at: https://storybook.js.org/docs/writing-stories
+const meta: any = {
+    title: 'sequence/configuration/phase/ResponsePhaseConfiguration',
+    component: ResponsePhaseConfiguration,
+    args: {
+        "onUpdate:responsePhaseConfig": fn(),
+        studentGiveExplanation: false,
+        explanationMandatory: false
+    },
+    tags: ['autodocs', 'organisms'],
+    parameters: {
+        docs: {
+            description: {
+                story: 'The configuration of the Response phase.'
+            }
+        }
+    }
+} satisfies Meta<typeof ResponsePhaseConfiguration>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+
+export const Primary: Story = {
+    parameters: {
+        docs: {
+            description: {
+                story: 'Default configuration of the response phase.'
+            }
+        }
+    }
+};
+
+export const ExplanationMandatory: Story = {
+    parameters: {
+        docs: {
+            description: {
+                story: 'The student must provide an explanation for their answer.'
+            }
+        }
+    },
+    args: {
+        explanationMandatory: true
+    }
+};
