+++ ui-components/src/stories/sequence/configuration/phase/ResultPhaseConfiguration.stories.ts
@@ -1,0 +1,57 @@
+/*
+ * Elaastic - formative assessment system
+ * Copyright (C) 2019. University Toulouse 1 Capitole, University Toulouse 3 Paul Sabatier
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+import type { Meta, StoryObj } from '@storybook/vue3-vite'
+import ResultPhaseConfiguration from '@/components/sequence/configuration/phase/ResultPhaseConfiguration.vue'
+import { fn } from 'storybook/test'
+
+// More on how to set up stories at: https://storybook.js.org/docs/writing-stories
+const meta: any = {
+  title: 'sequence/configuration/phase/ResultPhaseConfiguration',
+  component: ResultPhaseConfiguration,
+  args: {
+    'onUpdate:evaluationByIa': fn(),
+    aiIsActivated: true,
+    evaluationByIa: false,
+  },
+  tags: ['autodocs', 'organisms'],
+  parameters: {
+    docs: {
+      description: {
+        story: 'TODO',
+      },
+    },
+  },
+} satisfies Meta<typeof ResultPhaseConfiguration>
+
+export default meta
+type Story = StoryObj<typeof meta>
+
+export const Primary: Story = {
+  args: {
+    evaluationByIa: false,
+  },
+
+  parameters: {
+    docs: {
+      description: {
+        story: 'TODO',
+      },
+    },
+  },
+}
